Added support for Android data type double

+++ src/main/java/com/marvik/apis/dbcrudgen/core/databases/sqlite/SQLiteUtils.java
@@ -72,7 +72,7 @@ public class SQLiteUtils {
             return "FLOAT";
         } // Float
         if (mysqlDataType.equalsIgnoreCase("Decimal")) {
-            return "FLOAT";
+            return "DOUBLE";
         }
         // Long
         if (mysqlDataType.equalsIgnoreCase("Long")) {
